+++ src/bitcoin-chainstate.cpp
@@ -15,7 +15,7 @@
 #include <kernel/context.h>
 #include <kernel/validation_cache_sizes.h>
 
-#include <chainparams.h>
+#include <kernel/chainparams.h>
 #include <consensus/validation.h>
 #include <core_io.h>
 #include <node/blockstorage.h>
@@ -51,8 +51,9 @@ int main(int argc, char* argv[])
 
 
     // SETUP: Misc Globals
-    SelectParams(CBaseChainParams::MAIN);
-    const CChainParams& chainparams = Params();
+    SetGlobalBaseParams(CBaseChainParams::MAIN);
+    auto chainparams_ptr = CChainParams::Main();
+    const CChainParams& chainparams = *chainparams_ptr;
 
     kernel::Context kernel_context{};
     // We can't use a goto here, but we can use an assert since none of the
