+++ src/chainparams.h
@@ -6,6 +6,8 @@
 #ifndef BITCOIN_CHAINPARAMS_H
 #define BITCOIN_CHAINPARAMS_H
 
+#include <kernel/chainparams.h>
+
 #include <chainparamsbase.h>
 #include <consensus/params.h>
 #include <cstdint>
@@ -18,190 +20,6 @@
 #include <memory>
 #include <string>
 #include <vector>
-
-typedef std::map<int, uint256> MapCheckpoints;
-
-struct CCheckpointData {
-    MapCheckpoints mapCheckpoints;
-
-    int GetHeight() const {
-        const auto& final_checkpoint = mapCheckpoints.rbegin();
-        return final_checkpoint->first /* height */;
-    }
-};
-
-struct AssumeutxoHash : public BaseHash<uint256> {
-    explicit AssumeutxoHash(const uint256& hash) : BaseHash(hash) {}
-};
-
-/**
- * Holds configuration for use during UTXO snapshot load and validation. The contents
- * here are security critical, since they dictate which UTXO snapshots are recognized
- * as valid.
- */
-struct AssumeutxoData {
-    //! The expected hash of the deserialized UTXO set.
-    const AssumeutxoHash hash_serialized;
-
-    //! Used to populate the nChainTx value, which is used during BlockManager::LoadBlockIndex().
-    //!
-    //! We need to hardcode the value here because this is computed cumulatively using block data,
-    //! which we do not necessarily have at the time of snapshot load.
-    const unsigned int nChainTx;
-};
-
-using MapAssumeutxo = std::map<int, const AssumeutxoData>;
-
-/**
- * Holds various statistics on transactions within a chain. Used to estimate
- * verification progress during chain sync.
- *
- * See also: CChainParams::TxData, GuessVerificationProgress.
- */
-struct ChainTxData {
-    int64_t nTime;    //!< UNIX timestamp of last known number of transactions
-    int64_t nTxCount; //!< total number of transactions between genesis and that timestamp
-    double dTxRate;   //!< estimated number of transactions per second after that timestamp
-};
-
-/**
- * CChainParams defines various tweakable parameters of a given instance of the
- * Bitcoin system.
- */
-class CChainParams
-{
-public:
-    enum Base58Type {
-        PUBKEY_ADDRESS,
-        SCRIPT_ADDRESS,
-        SECRET_KEY,
-        EXT_PUBLIC_KEY,
-        EXT_SECRET_KEY,
-
-        MAX_BASE58_TYPES
-    };
-
-    const Consensus::Params& GetConsensus() const { return consensus; }
-    const CMessageHeader::MessageStartChars& MessageStart() const { return pchMessageStart; }
-    uint16_t GetDefaultPort() const { return nDefaultPort; }
-    uint16_t GetDefaultPort(Network net) const
-    {
-        return net == NET_I2P ? I2P_SAM31_PORT : GetDefaultPort();
-    }
-    uint16_t GetDefaultPort(const std::string& addr) const
-    {
-        CNetAddr a;
-        return a.SetSpecial(addr) ? GetDefaultPort(a.GetNetwork()) : GetDefaultPort();
-    }
-
-    const CBlock& GenesisBlock() const { return genesis; }
-    /** Default value for -checkmempool and -checkblockindex argument */
-    bool DefaultConsistencyChecks() const { return fDefaultConsistencyChecks; }
-    /** Policy: Filter transactions that do not match well-defined patterns */
-    bool RequireStandard() const { return fRequireStandard; }
-    /** If this chain is exclusively used for testing */
-    bool IsTestChain() const { return m_is_test_chain; }
-    /** If this chain allows time to be mocked */
-    bool IsMockableChain() const { return m_is_mockable_chain; }
-    uint64_t PruneAfterHeight() const { return nPruneAfterHeight; }
-    /** Minimum free space (in GB) needed for data directory */
-    uint64_t AssumedBlockchainSize() const { return m_assumed_blockchain_size; }
-    /** Minimum free space (in GB) needed for data directory when pruned; Does not include prune target*/
-    uint64_t AssumedChainStateSize() const { return m_assumed_chain_state_size; }
-    /** Whether it is possible to mine blocks on demand (no retargeting) */
-    bool MineBlocksOnDemand() const { return consensus.fPowNoRetargeting; }
-    /** Return the network string */
-    std::string NetworkIDString() const { return strNetworkID; }
-    /** Return the list of hostnames to look up for DNS seeds */
-    const std::vector<std::string>& DNSSeeds() const { return vSeeds; }
-    const std::vector<unsigned char>& Base58Prefix(Base58Type type) const { return base58Prefixes[type]; }
-    const std::string& Bech32HRP() const { return bech32_hrp; }
-    const std::vector<uint8_t>& FixedSeeds() const { return vFixedSeeds; }
-    const CCheckpointData& Checkpoints() const { return checkpointData; }
-
-    //! Get allowed assumeutxo configuration.
-    //! @see ChainstateManager
-    const MapAssumeutxo& Assumeutxo() const { return m_assumeutxo_data; }
-
-    const ChainTxData& TxData() const { return chainTxData; }
-
-    /**
-     * SigNetOptions holds configurations for creating a signet CChainParams.
-     */
-    struct SigNetOptions {
-        std::vector<uint8_t> challenge;
-        std::vector<std::string> seeds = {
-            "seed.signet.bitcoin.sprovoost.nl.",
-            "178.128.221.177",
-            "v7ajjeirttkbnt32wpy3c6w3emwnfr3fkla7hpxcfokr3ysd3kqtzmqd.onion:38333"
-        };
-    };
-
-    /**
-     * VersionBitsParameters holds activation parameters
-     */
-    struct VersionBitsParameters {
-        int64_t start_time;
-        int64_t timeout;
-        int min_activation_height;
-    };
-
-    /**
-     * RegTestOptions holds configurations for creating a regtest CChainParams.
-     */
-    struct RegTestOptions {
-        std::unordered_map<Consensus::DeploymentPos, VersionBitsParameters> version_bits_parameters{};
-        std::unordered_map<Consensus::BuriedDeployment, int> activation_heights{};
-        bool fastprune{false};
-    };
-
-    /**
-     * Creates and returns a std::unique_ptr<CChainParams> for regtest.
-     * @return std::unique_ptr<const CChainParams>
-     */
-    static std::unique_ptr<const CChainParams> RegTest(const RegTestOptions& options);
-
-    /**
-     * Creates and returns a std::unique_ptr<CChainParams> for signet.
-     * @return std::unique_ptr<const CChainParams>
-     */
-    static std::unique_ptr<const CChainParams> SigNet(const SigNetOptions& options);
-
-    /**
-     * Creates and returns a std::unique_ptr<CChainParams> for mainnet.
-     * @return std::unique_ptr<const CChainParams>
-     */
-    static std::unique_ptr<const CChainParams> Main();
-
-    /**
-     * Creates and returns a std::unique_ptr<CChainParams> for testnet.
-     * @return std::unique_ptr<const CChainParams>
-     */
-    static std::unique_ptr<const CChainParams> TestNet();
-
-protected:
-    CChainParams() {}
-
-    Consensus::Params consensus;
-    CMessageHeader::MessageStartChars pchMessageStart;
-    uint16_t nDefaultPort;
-    uint64_t nPruneAfterHeight;
-    uint64_t m_assumed_blockchain_size;
-    uint64_t m_assumed_chain_state_size;
-    std::vector<std::string> vSeeds;
-    std::vector<unsigned char> base58Prefixes[MAX_BASE58_TYPES];
-    std::string bech32_hrp;
-    std::string strNetworkID;
-    CBlock genesis;
-    std::vector<uint8_t> vFixedSeeds;
-    bool fDefaultConsistencyChecks;
-    bool fRequireStandard;
-    bool m_is_test_chain;
-    bool m_is_mockable_chain;
-    CCheckpointData checkpointData;
-    MapAssumeutxo m_assumeutxo_data;
-    ChainTxData chainTxData;
-};
 
 /**
  * Creates and returns a std::unique_ptr<CChainParams> of the chosen chain.
